auth-service/requirements.txt
aiohttp==3.10.5
alembic-postgresql-enum==0.4.0
alembic==1.13.2
async_fastapi_jwt_auth==0.6.6
backoff==2.2.1
fastapi==0.111.0
opentelemetry-api==1.27.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-sdk==1.27.0
psycopg[binary]==3.1.19
pydantic_settings==2.6.0
redis==5.0.8
sentry-sdk[fastapi]==2.17.0
sqlalchemy[asyncio]==2.0.32
typer==0.12.4
